requirements.txt
flask==3.0.3
pydantic==2.7.0
python-dotenv==1.0.1
openai==1.30.1 